requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
openai==1.3.5
azure-identity==1.15.0
SQLAlchemy==2.0.20
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
Flask-Cors==4.0.0
